// packages/spark-bff/src/generated/typings/numberReservationService.ts
/**
 * This file was auto-generated by openapi-typescript.
 * Do not make direct changes to the file.
 */

export interface paths {
  '/v1/shop/numbers/mobile': {
    get: operations['getAvailableMobileNumbers'];
  };
  '/v1/shop/numbers/mobile/reserve': {
    post: operations['reserveMobileNumber'];
  };
  '/v1/shop/numbers/mobile/portIn': {
    post: operations['portInMobileNumber'];
  };
}

export interface definitions {
  MobileNumbersResponse: definitions['Response'] & {
    /** The reservation Id to be used for this transaction. */
    reservationId: string;
    /** The paging info provided by downstream for further calls on the same reservation id */
    paging: string;
    /** list of available numbers. */
    lineNumbers?: string[];
  };
  MobileNumberReservationRequest: {
    /** The reservation id retrieved when this number was fetched as part of availability. */
    reservationId: string;
    /** The mobile number to be reserved. */
    lineNumber: string;
  };
  MobileNumberPortInRequest: {
    /** The mobile number to be ported. */
    lineNumber: string;
    /** The current valid service provider id. */
    serviceProviderId: string;
  };
  MobileNumberPortInResponse: definitions['Response'] & {
    /** The reservation id if the portin is success. */
    reservationId: string;
  };
  Response: {
    /** List of Messages */
    messages: definitions['Message'][];
  };
  Message: {
    message: string;
    code: number;
  };
}

export interface operations {
  getAvailableMobileNumbers: {
    parameters: {
      query: {
        /** Indicates the expected number of available mobile numbers. */
        count?: number;
        /** Provided to indicate the paging when more numbers are requested against a reservationId */
        pagingId?: string;
        /** Provided when more numbers are required against a reservationId. */
        reservationId?: string;
      };
    };
    responses: {
      /** Successfully retrieved mobile numbers. */
      200: {
        schema: definitions['MobileNumbersResponse'];
      };
      /** Bad Request */
      400: {
        schema: definitions['Response'];
      };
      /** Internal Server Error */
      500: {
        schema: definitions['Response'];
      };
    };
  };
  reserveMobileNumber: {
    parameters: {
      query: {
        /** The account number against which the reservation has to happen. */
        accountNumber?: string;
      };
      body: {
        /** Request holding mobile number reservation details. */
        mobileNumberReservationRequest: definitions['MobileNumberReservationRequest'];
      };
    };
    responses: {
      /** Successfully reserved the mobile number. */
      200: {
        schema: definitions['Response'];
      };
      /** Bad Request */
      400: {
        schema: definitions['Response'];
      };
      /** The provided token is invalid. */
      401: {
        schema: definitions['Response'];
      };
      /** Internal Server Error */
      500: {
        schema: definitions['Response'];
      };
    };
  };
  portInMobileNumber: {
    parameters: {
      body: {
        /** Request holding mobile number port in details. */
        mobileNumberPortInRequest: definitions['MobileNumberPortInRequest'];
      };
    };
    responses: {
      /** Successfully ported the mobile number. */
      200: {
        schema: definitions['MobileNumberPortInResponse'];
      };
      /** Bad Request */
      400: {
        schema: definitions['Response'];
      };
      /** The provided token is invalid. */
      401: {
        schema: definitions['Response'];
      };
      /** Internal Server Error */
      500: {
        schema: definitions['Response'];
      };
    };
  };
}